[MCCAS] Split debug info section
This commit changes the MCCASBuilder so that it is able to split the
debug info section based on its compilation units. For each compilation
unit in the debug info section, one CAS object is created.

This patch does not handle dwarf 64 formats and does not attempt to
enable deduplication of debug info sections.

Testing is done with the verify mode of the CAS backend, which compares
an object file generated through CAS with an object file generated
through the traditional flow. Once tooling is developed, we should be
able to check that multiple CAS objects are actually created.