Make the compilation cache key and the CAS outputs data output path independent
Instead of storing CAS outputs by the specific path that was used at the time of invocation, associate them with an "output kind name" so that they are re-usable even with different output paths.

The practical benefit of this is that we can get cache hits when source files are part of multiple products (e.g. when source files are included in multiple Xcode test targets) or when PCH files end up with the same caching key and output but the build system is not in a position identify this.