[MCCAS] Introduce llvm-cas-mc-dump tool
This is an initial attempt at a tool that navigates the CAS tree of an
object using the MC Schema and prints the tree structure.

Currently, it only prints the node kind, followed by its ID, followed by
a nested list of children nodes (recursively). Since the CAS is always a
DAG, there is no risk of infinite loops here.

For example, one possible output:

```
mc:assembler    llvmcas://5f77...
  mc:header       llvmcas://5264...
  mc:group        llvmcas://dfa25...
    mc:section      llvmcas://1a762...
      mc:debug_info_cu llvmcas://64c13...
      mc:debug_info_cu llvmcas://d0c6c744...
```

Since the first use case for this tool is to check how many objects each debug
section contains, we provide a flag to filter non-debug sections.

There are limitations to the tool at this point, all of which can be resolved
in the future:

1. It does not understand the _contents_ of each CAS object beyond the "type"
of object it is. So, for example, we cannot print the name of a section.

2. If one object contains multiple edges with the same destination, the MC CAS
encoding may choose to only add unique edges, and encode the duplicates inside
the "Data" section of the object. The tool currently does not look at those, so
there may be situations in which duplicate edges are not printed.

3. Because we don't use any nesting marks other than whitespaces, and
FileCheck does not usually distinguish whitespaces, nesting levels may
be ambiguous for FileCheck. We should explore using `{}` in the future.