[MCCAS][nfc] Refactor Dwarf Section handling
Now that there is more clarity on how Dwarf sections are going to be
handled, this commit refactors the existing code into a more uniform
approach for handling these sections.

1. All the queries to the MCAssembler that attempt to find debug
information are only done once and stored into a cache object for ease
of access.

2. The main loop that iterates over all sections now dispatches the
handling of debug sections to specialized functions, removing any
handling that was previously inlined into the loop.

3. The specialized functions above now have a more consistent naming and
documentation schemes.

This is not the full extent of possible refactors, but it provides a
better starting point for the upcoming work, which will split more
complicated sections.