[MCCAS][nfc] Refactor Dwarf Debug Info Splitting loop
This commit refactors the main Dwarf Debug Info splitting loop so that
it creates a sequence of Mutable Array Refs, instead of having a single
loop that both splits the data and consumes it.

This will make it easier to split abbreviation sections and mutate the
contents of the debug info section in subsequent commits.