[include-tree] Enable producing an `include-tree` via a driver invocation and replaying it for compilation
The `include-tree` is added as an alternative method to `casfs` for dep-scanning and for driving a compilation.
For a `clang-cache` invocation it is enabled via the `CLANG_CACHE_ENABLE_INCLUDE_TREE` environment variable.

Due to avoiding header file lookups there's a measurable improvement for preprocessing performance.
The following are wall-time measurements on an M1Pro with a thin-LTO enabled build; preprocess-only compilations
are performed on all the files that are linked to the clang executable, using:

* The arguments produced with include-tree dep-scanning
* The arguments produced with CASFS dep-scanning
* The normal compilation arguments

|              | casfs | normal |
|--------------|-------|--------|
| include-tree | -7.2% | -21.5% |

(cherry picked from commit 5f899118c356b54d11a2bd448ddabf6f2a0346b9)