[driver] Ensure that we save & replay the serialized diagnostics file for caching purposes
Previously the diagnostics from the compilation would be missing from the produced serialized diagnostics file.
There remains a small issue that the "compile job cache" remarks are not written out to the serialized diagnostics file but
I consider this a minor issue compared to missing all of the compilation diagnostics.

However the way we handle diagnostics for caching is still sub-optimal, I extended a `FIXME` with a potential approach
to improve the handling.