[clang][cas] Remove -MT and -MP from the cache key
Pull the dependency options that only affect the output but not the list
of filenames out of the cache key, and apply them instead when replaying
the list of dependencies from the cache. The dependency file target
name(s) are typically an output path, so we want to avoid it showing up
in the cache key.

For now, the options that *do* affect the list of dependencies like
the difference between -MMD and -MD are still part of the key so that we
do not store many unnecessary system paths if they will not be used. It
would be straightforward to change that if it turned out to be
worthwhile by storing flags like IsSystem, IsModuleFile, etc.

(cherry picked from commit 5e98b81f020e105986fa397fe09b20c5823fc38b)
(cherry picked from commit cbab7a353b2a361e0ad6e7dfb5d2f916b37e4e82)