[MCCAS] Add DebugInfoSectionRef class and verification methods
This commit introduces a new CAS object type: DebugInfoSectionRef. This
is meant to replace SectionInfoRefs for the `__debug_section` section.
As a consequence, distinguishing these sections in the code can be done
with a `Cast` operation, as opposed to having to load and check their
child CAS objects.

Furthermore, verification code is provided so that all error checking is
done in a single abstraction level.