[MCCAS][nfc] Add writing capabilities to debug CU splitting function
Subsequent patches will need to overwrite the abbreviation offset when
splitting the CUs and when materialization them. In preparation for
that, this patch adds writing capabilities to the function that handles
abbreviation offsets in CUs. Extracting the offset itself is not changed
by this patch, such functionality was already there prior to the commit.

In addition, a forward declaration is added so that other parts of the
file can access this function (while minimizing the diff). A subsequent
NFC commit should move the function definition to the top of the file.