[MCCAS] Zero out abbreviation offsets in CURefs
This commit introduces zeroing out of the abbreviation offsets in
CURefs.

Some changes are required in the CASDWARFObject classes. In particular,
we choose not to overwrite the false "0x0000" offset printed when we
dump the CU header. This is done for simplicity: from the dump utility's
point of view, the real offset is _zero_, since we only give it the
abbreviation fragment it needs, not the fully assembled abbreviation
section. For verification purposes, we print an extra line with the true
offset of the abbreviation section.