[lldb] Implement cache for swift typeref metadata
Implement a mechanism to cache field descriptor location information.
One of the main problems of using type metadata as debug info, when in
comes to field descriptors in particular, is that there's no easy way
to find the correct type given a (string key), so a linear search over
every source of reflection information must be done. This patch
implements a cache which allows for field descriptors to be found in a
more efficient manner.