[Basic] Always return a FileEntry*-containing FileEntryRef
It's possible to be in a situation where the `FileEntryRef` returned by
`FileManager::getFileRef` contains a `MapEntry *` rather than
`FileEntry *`. All callers expect the latter, so on non-assert builds
(where the `PointerUnion` conversion doesn't fail) this *at best* causes
a crash.

To see how this can happen, consider the following overlays:
  - B -> C
  - A -> B
  - Both B and C exist on the real filesystem

Looking up B and then A will result in `SeenFileEntries` containing:
  - B -> `MapEntry *` (which points to C)
  - C -> `FileEntry *`
  - A -> `MapEntry *` (which points to B)

The return for A will now be a `MapEntry *` rather than a `FileEntry *`.

As a temporary workaround, always grab the base `FileEntryRef` in these
cases. ToT has already fixed this case by virtue of `FileEntryRef` now
*expecting* redirections in order to allow returning the requested path.

Resolves rdar://82356878.