Sink SwiftHost.cpp into HostInfo (NFC)
This refactors the code that currently lives in Expression/SwiftHost
to find a new home in HostInfo. The idea is to eventually allow
different implementations per host platform. Note that this is
intentionally just moving the code around, there has been no attempt
at cleaning up the code or interface.

rdar://92640609