This patch changes how we store compile unit information in the CAS. In an effort to get deduplication in the blocks that represent one compile unit, we have moved any information that will not deduplicate into its own CAS object denoted by the name DebugInfoFixupsRef. The information that will deduplicate will be stored in the DebugInfoCURef. We make the distinction between this by separating out DW_FORM_strp and DW_FORM_sec_offset into the DebugInfoFixupsRef. When materializing the CAS, we iterate over the abbreviations to re-create a compile unit as it would be in the object file.
This change also edits llvm-cas-dump to print Compile Units properly. A DebugInfoCURef doesn't contain all the information in a Compile Unit, but only the information that deduplicates, therefore the dwarfdump needs to be changed to get the full Compile Unit data to produce an accurate output.

This change also adds --show-form and --v options to llvm-cas-dump which do the same as -F and -v in dwarfdump.

This change also adds a test to check for debug information deduplication.