[clang][cas] Create node schema for comilation cache result
Factor the code for creating and reading a compilation result out of
cc1_main and create a CompileJobCacheResult proxy for working with
cached compilation results. This is in preparation for accessing cached
module compilations.

Also, make it a hard error to create an output file of unknown kind -
previously these could create cache values that depended on output paths
not in the cache key.