[clang][cas/include-tree] Keep track of the filenames as they were recorded in the PCH
The filenames as recorded in the PCH may differ from the filenames as headers are included for the main file.
Keep track of the PCH filenames so that we can record both variants for a `FileEntry` in the include-tree.

(cherry picked from commit 296b4da4ad7bfdc2b7ae4bdc8b361bf9f1a5b998)