[clang][cas] Ignore the `-fdepfile-entry=` dependency option when calculating the cache key
`-fdepfile-entry=` is a dependency option that can be treated as "transparent", not affecting the
cached compilation. The passed-in filenames are added at the point where the dependency file is generated
and are not read from the file system.

rdar://99683427
(cherry picked from commit dde9a72957d7964532affa8ee1118e53576bc99b)