[clang-scan-deps][cas] Initial support for caching modules using casfs
When a CAS and ActionCache are available to to the scanner, we can
produce build commands for modules as well as the main TU that use
the cc1 option -fcache-compile-job to enable compilation caching. This
makes use of the newly added -fmodule-file-cache-key= to support
creating commands that import a given module via an action cache key
without needing that module to have been built yet. This means that
unlike with PCH, we do not need to scan commands that use modules in
dependency order with respect to building those modules, and we do not
need to read the pcm files back from disk repeatedly in order to ingest
them into a casfs. Like any scanned explicit modules build, the modules
must still be *built* in dependency order by the build system, or else
the action cache lookup will fail in the same places that an uncached
build would fail due to a missing .pcm on disk.

This commit contains the primary changes needed to build modules files
via the cache. However, there are some additional changes that will be
needed before this is practical to use:

* Support for mixing PCH + Modules
* Support prefix mapping for paths
* Thinning out the set of files ingested into the casfs to reduce
  spurious cache misses - in particular, removing unnecessary volatile
  files such as the implicit module pcm files created by the scanner.