Revert "Revert "Be more careful to maintain quoting information when parsing commands.""
    This reverts commit ac05bc0524c66c74278b26742896a4c634c034cf.

I had incorrectly removed one set of checks in the option handling in
Options::ParseAlias because I couldn't see what it is for.  It was a
bit obscure, but it handled the case where you pass "-something=other --"
as the input_line, which caused the built-in "run" alias not to return
the right value for IsDashDashCommand, causing TestHelp.py to fail.

(cherry picked from commit f3d0bda5344e8429121cb50aba5502c1b399e5cd)