[clang-cache] Introduce `CLANG_CACHE_TEST_DETERMINISTIC_OUTPUTS` environment variable
This is used for testing purposes. It instructs `clang-cache` to run the compilation twice, without replaying,
to check that we get the same compilation artifacts for the same key. If they are not the same the action cache
will trigger a fatal error.

(cherry picked from commit 814a3feb586ab24e39e3bbecc10e30b051ba04e1)