[clang][cas/remote] Introduce a mechanism for storing & retrieving compilation artifacts using an external process service
Provide a mechanism for clang to connect to an external process that can store and retrieve compilation artifacts,
associated with a particular compilation key. How the external process caches the artifacts is left as implementation detail
but the primary motivation is to enable connecting with custom processes that can store and retrieve compilation artifacts via
a network distributed cache.

The pieces are:

* `gRPC` service protocol definitions used as the communication API with the external process:
  * API for a key-value service
  * API for a CAS-like service
* a `llvm-remote-cache-test` tool that implements a server for the remote caching protocol, using an on-disk `cas::ObjectStore`
  and `cas::ActionCache`, intended for testing purposes.
* a remote cache client API for connecting to the external process
* in `cc1_main`, an implementation of `CachingOutputs` that uses the client API for storing and retrieving compilation outputs from the external service

The external service is only used for caching the compilation outputs; dependency scanning, compilation key computation, and the input
to the caching compilation is still processed using the `cas::ObjectStore`. Using the local on-disk `cas::ObjectStore` for the inputs ensures
efficient processing for the clang caching implementation.

The remote cache mechanism requires `gRPC` to available, and that the CMake variable `LLVM_CAS_ENABLE_REMOTE_CACHE` is enabled.

It is enabled when using the `clang-cache` launcher by setting `LLVM_CACHE_REMOTE_SERVICE_SOCKET_PATH` environment variable to the
unix domain socket path that the external service is listening on.

(cherry picked from commit e4a97f1e27631d09870531ab88a37ce21da672fa)