[clang][cas] Allow spelling differences in paths for DependencyScanningCASFilesystem
Do not require getDirectiveTokens to receive an identical filename to a
previous filesystem operation, as this is hard to guarantee when the
calls are far apart and go through complex filesystem layers such as
FileManager and VFS, any of which could modify the path. In practice,
this was causing breakage when using a VFS which canonicalized away "./"
in a path.

Note: because all the actual filesystem accesses go through
CachingOnDiskFileSystem, and the directive scanning itself is cached in
the ActionCache, the cost of mismatching the spelling should be low.

(cherry picked from commit 3f96618a56383c56806d2ff668d9454580c0e7c2)
(cherry picked from commit 1854bdc5c47e3ae0d100d66c84083be985605855)