[lldb] Add newline in output of `target modules lookup`
This adds a line break between each result address in the output of the
lldb command `target modules lookup`. Before this change, a new address
result will be printed on the same line as the summary of the last
result, making the output difficult to view.

Also adds a test for this command.

Reviewed By: labath

Differential Revision: https://reviews.llvm.org/D134111

(cherry picked from commit a426753ef0290362a0268f917dba46d6b9744b02)