Allow DynamicLoaderDarwinKernel to activate without binary
In https://reviews.llvm.org/D133534 I made a little cleanup
to DynamicLoaderDarwinKernel::CreateInstance and unintentionally
changed the logic.  Previously it would not create an instance
if there was a binary given to lldb and it was not a kernel.
With my change, the absence of any binary would also cause it
to not create.  So connecting to a kernel without any binaries
would fail.

rdar://100985097
(cherry picked from commit b8a8c2d47a38ba08012fe9cbb28169a1e0f7be2a)