[lldb] Implement extra inhabitants calculation for foreign types in LLDB
The Swift reflection machinery uses the number of extra inhabitants to
figure out if enum cases were compacted inside the payload. The LLDB
external provider was always setting that value to 0, which would
subsequently make reflection think that a field was bigger than it
actually was, and causing accesses to subsequent fields to be incorrect.
This patch implements returning the correct number of extra inhabitants
in the external provider.

rdar://100708229