[RemoteCachingService] Remove the initial connection timeout
If the execution environment gets into a state where the processes don't make progress, for a period longer than the timeout,
it can cause unnecessary compilation errors if the connection fails due to the timeout. Note that the connection still fails
if the socket path does not exist or there is no process listening on it.

rdar://101772572