[MCCAS] Introduce debug info CAS blocks for Tag-splitting impl
This commit introduces a few CAS block types to aid with the
implementation of the new Tag-splitting approach for debug information:

0. DIEAbbrevRef, to hold the abbreviation contents of a Tag. This
corresponds to a single abbreviation entry in Dwarf.
1. DIEAbbrevSetRef, to hold sets of abbreviation blocks and enable
efficient references into it. This corresponds to the entire
abbreviation section of Dwarf.
2. DIEDistinctDataRef, to hold the distinct data that is not expected to
deduplicate across translation units.
3. DIEDataRef, to hold the data expected to deduplicate in a (collection
of) Tag(s). This block may also store references to other DIEDataRefs;
as such, we cannot use the "SIMPLE_DATA" or "SIMPLE_GROUP" macros.
4. A DIETopLevelRef block, to hold three references, one for each of
blocks 1-3 above. This is done for convenience of materialization.

Note that some of these blocks are redundant with the existing
implementation. However, to ease with the transition and to allow for
smaller implementation steps, we choose to add new blocks types.