[lldb] Fix Array<T>.SubSequence synthetic (#5523)
`Array<T>.SubSequence` is a typealias to `ArraySlice<T>`. LLDB provides a synthetic type for `ArraySlice`. When printing a value of `Array<T>.SubSequence`, the formatting system correctly peels off the typealias and selects the `ArraySlice` synthetic provider.

However the implementation of the `ArraySlice` synthetic provider (see `SwiftArrayBufferHandler`) handles a number of related array types, including `Array`, `ContiguousArray`, `ArraySlice`, and more.

In handling these types, `SwiftArrayBufferHandler` does string matching on the type name to determine its behavior. This string matching is done on the exact type, no typealias layers are removed.

For `Array<T>.SubSequence` to be handled as `ArraySlice`, the string matching needs to be updated accordingly. This change does that.

rdar://92898538